+++ src/pages/Question/Post/AnswerPost.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 
 import MDEditor from "@uiw/react-md-editor";
 
@@ -14,6 +14,8 @@ import { useSessionContext } from "../../../contexts/SessionContext";
 import styles from "./Post.module.scss";
 import { api } from "../../../api/api";
 import { toast } from "react-toastify";
+import axios from "axios";
+import { confirmAlert } from "react-confirm-alert";
 
 interface PostProps {
   answer: Answer;
@@ -32,15 +34,40 @@ const AnswerPost: React.FC<PostProps> = ({
   const auth = userInfo?.id === answer.user.id;
   const [onAdd, setOnAdd] = useState<boolean>(false);
   const [comment, setComment] = useState<string>("");
+  const navigate = useNavigate();
 
-  const handleDelete = async () => {
-    try {
-      await api.deleteAnswer(answer.id);
-      setReset(!reset);
-      // navigate(`/questions/${questionId}`);
-    } catch (err) {
-      console.error(err);
-    }
+  const handleDelete = () => {
+    confirmAlert({
+      title: "Confirm",
+      message: "Are you sure?",
+      buttons: [
+        {
+          label: "Yes",
+          onClick: async () => {
+            try {
+              await api.deleteAnswer(answer.id);
+              setReset(!reset);
+              toast.info("Answer deleted!");
+            } catch (err) {
+              if (axios.isAxiosError(err) && err.response) {
+                if (err.response.status === 401) {
+                  toast.error("Please sign in first!");
+                  navigate("/singin");
+                } else if (err.response.status === 403) {
+                  toast.error("Cannot delete other user's answer");
+                } else if (err.response.status === 404) {
+                  toast.error("The answer does not exist");
+                } else console.error(err.response.data);
+              } else console.error(err);
+            }
+          },
+        },
+        {
+          label: "Cancel",
+          onClick: () => {},
+        },
+      ],
+    });
   };
 
   const handleCommentSubmit: React.FormEventHandler<HTMLFormElement> = async (
@@ -53,12 +80,21 @@ const AnswerPost: React.FC<PostProps> = ({
         return;
       }
       await api.postAnswerComment(answer.id, comment);
-      // navigate(`/questions/${questionId}`);
       setReset(!reset);
       setOnAdd(false);
       setComment("");
+      toast.info("Comment created!");
     } catch (err) {
-      console.error(err);
+      if (axios.isAxiosError(err) && err.response) {
+        if (err.response.status === 401) {
+          toast.error("Please sign in first!");
+          navigate("/signin");
+        } else if (err.response.status === 404) {
+          toast.error("The answer does not exist");
+        } else if (err.response.status === 405) {
+          toast.error("Invalid comment content");
+        } else console.error(err.response.data);
+      } else console.error(err);
     }
   };
 
@@ -112,7 +148,7 @@ const AnswerPost: React.FC<PostProps> = ({
           <div className={styles.activityContainer}>
             <UserCard
               user={answer.user}
-              timestamp="2021-11-26 20:45:00Z"
+              date={new Date(answer.createdAt + "Z")}
               isQuestion={false}
               isEdited={false}
             />
